+++ src/lib/supabase/database.types.ts
@@ -22,16 +22,19 @@ export interface Database {
         Row: {
           id: string;
           email: string | null;
+          buyer_name_info: string;
           created_at: string;
         };
         Insert: {
           id: string;
           email?: string | null;
+          buyer_name_info?: string;
           created_at?: string;
         };
         Update: {
           id?: string;
           email?: string | null;
+          buyer_name_info?: string;
           created_at?: string;
         };
         Relationships: [];